restructured files and added sqs implementation
